Add whitelist_python.py (#30)
* Add whitelist_python.py

A standard whitelist file for Python/stdlib.

Closes #21

* Add testcases for whitelist_python

Add a testcase to check whether or not is the whitelist_python.py
file is being executed normally and another one to check that code
containing parts from whitelist_python is really being marked as
used or not.

* Add more test cases for test_whitelist_python

Add a test which would check wether or not there is an error raised
when example_test file did not give an error despite of having some
dead code (This will generally happen because of wrong configuration
of tests, eg., when I mistakenly ran the test cases with python
executable rather than vulture, example file obviously ran fine. :) )

* Rename tests and alphabetical imports

* test whitelist_python by calling vulture on it

Previousyly, we were testing our whitelist file by calling vulture on
an example file and supplying whitelist_python.py. Now, we would call
vulture on whitelist_file for the same.