Treat `getattr/hasattr(obj, "constant_string", ...)` as a reference to `obj.constant_string`
I realize handling `getattr` is a slippery slope. I think this case makes sense and is a reasonable place to draw the line.

My rationale is that the following code does the same thing, counts as a reference to obj.constant_string, but is not what most people would write.
```
try:
    x = obj.constant_string
except AttributeError:
    x = None
```