Don't override TOML paths with empty CLI paths.
The paths CLI argument did not use the default=missing sentinel, and so
returned an empty list if no paths were given to the CLI. This meant any
paths configured in pyproject.toml were overridden with an empty list,
making the path= setting useless.

Adding the sentinel to the CLI fixes this behaviour. A regression test
is also added to make sure TOML paths are used if no CLI paths are
given.